add cell music

+++ src/main/java/com/example/data/repository/SongCategoryRepository.java
@@ -3,6 +3,10 @@ package com.example.data.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.data.entity.SongCategoryEntity;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
-public interface SongCategoryRepository extends JpaRepository<SongCategoryEntity, Long> {
+import java.util.List;
+
+public interface SongCategoryRepository extends JpaRepository<SongCategoryEntity, Integer> {
 }
